Moving forms to different locations
The new locations are for proper rendering.
