+++ app/forms/sipity/forms/ulra/work_submissions.rb
@@ -1,0 +1,9 @@
+module Sipity
+  module Forms
+    module Ulra
+      # Container module for WorkSubmissions within ULRA
+      module WorkSubmissions
+      end
+    end
+  end
+end
