+++ app/views/sipity/controllers/ulra/work_submissions/publisher_information.html.erb
